Updated Puppetry Sketch
Changed around the Puppetry manager to have a different socket management strategy. Now there are no socket services by default and agents have to make an HTTP request to have one created for them. This allows for stronger assumptions around the provenance of the socket service that streamlines a lot of the error checking.