REWORK PuppetBehvaior System
Another rework to the PuppetBehavior system that should streamline subclass implementation. Now all puppet actions are implemented as single functions or coroutines that take a PuppetCommand as an argument. These can be registered in the PuppetBehavior and it will handle calling for queueing everything internally. This also brings back a change of PuppetCommands being a struct rather than a nullable class. Now instead of a null PuppetCommand you should use PuppetCommand.IDLE. A few other small changes throughout to support GetInfo and GetState commands.